Move header to source directory
